fix: address review feedback (#4539)
- Delete duplicate `deposit_screen_title` from all 10 locale files and
  point `DepositScreen.kt`'s fallback toolbar at the existing
  `tx_overview_screen_tx_deposit` ("Deposit") so non-LP/CACAO deposit
  flows stop rendering "Function Signature" as the screen title.
- Move `VerifyCardJsonDetails` from `ui/screens/swap` to `ui/components`
  so `ui/components/SignSolanaDisplayView.kt` stops importing across the
  components→screens layer boundary; flip the default `subtitleColor`
  from `alerts.success` to `alerts.info` since every current caller is a
  JSON/code dump with no success semantics.
- Defer `SolanaTransactionParser.parse` until first expansion in
  `SignSolanaDisplayView` (parse moved inside `AnimatedVisibility`) so
  non-expanding users never pay the JNI cost on the composition thread;
  switch the expansion state to `rememberSaveable` so it survives
  rotation/process death; drop the private `initiallyExpanded` parameter
  from `SignSolanaDisplayContent` since no production caller can seed it
  and previews can render the inner sections directly.
- Drop the `(programName)` suffix from the Solana instruction header so
  the row matches the browser extension and the instruction type no
  longer truncates at `maxLines = 1`.
- Switch `VerifyTransactionDetailsSection`'s click row from
  `indication = null` to `LocalIndication.current` so taps outside the
  chevron get a ripple.
- Restore the production-must-not-set-`true` KDoc on
  `initiallyExpandedDetails` in `VerifySendScreen`.
- Revert the unrelated "Data length" → "Data Length" casing tweak in
  the English `solana_accounts_data` plural.
- Use 函数签名 (matching the sibling `函数参数`) for the Chinese
  `verify_transaction_function_signature_title`.
