+++ app/src/main/res/values-de/strings.xml
@@ -214,7 +214,6 @@
     <string name="share_qr_screen_permission_required">Für Freigabevorgänge sind Dateiberechtigungen erforderlich.</string>
     <string name="token_selection_search_hint">Suchen…</string>
     <string name="verify_swap_agree_allowance">Ich bin mit der Bereitstellung eines unbegrenzten ERC20-Kontingents einverstanden</string>
-    <string name="deposit_screen_title">Funktionssignatur</string>
     <string name="deposit_screen_amount_title">Betrag</string>
     <string name="verify_deposit_memo_title">Memo</string>
     <string name="verify_deposit_node_address_title">Knotenadresse</string>
