+++ app/src/main/res/values-hr/strings.xml
@@ -214,7 +214,6 @@
     <string name="share_qr_screen_permission_required">Za sigurnosne kopije potrebno je dopuštenje datoteke.</string>
     <string name="token_selection_search_hint">Pretraživanje…</string>
     <string name="verify_swap_agree_allowance">Slažem se s pružanjem neograničenog ERC20 dodatka</string>
-    <string name="deposit_screen_title">Potpis funkcije</string>
     <string name="deposit_screen_amount_title">Iznos</string>
     <string name="verify_deposit_memo_title">dopis</string>
     <string name="verify_deposit_node_address_title">Adresa čvora</string>
