+++ app/src/main/res/values-ko/strings.xml
@@ -289,7 +289,6 @@
     <string name="share_qr_screen_permission_required">공유 작업을 위해 파일 권한이 필요합니다.</string>
     <string name="token_selection_search_hint">검색…</string>
     <string name="verify_swap_agree_allowance">ERC20 허용량 제공에 동의합니다</string>
-    <string name="deposit_screen_title">함수 서명</string>
     <string name="deposit_screen_amount_title">금액</string>
     <string name="verify_deposit_memo_title">메모</string>
     <string name="verify_deposit_node_address_title">노드 주소</string>
