+++ app/src/main/res/values-nl/strings.xml
@@ -213,7 +213,6 @@
     <string name="share_qr_screen_permission_required">Voor het delen van bewerkingen is bestandsmachtiging vereist.</string>
     <string name="token_selection_search_hint">Zoekopdracht…</string>
     <string name="verify_swap_agree_allowance">Ik ga akkoord met het verstrekken van een onbeperkte ERC20-toelage</string>
-    <string name="deposit_screen_title">Functiehandtekening</string>
     <string name="deposit_screen_amount_title">Bedrag</string>
     <string name="verify_deposit_memo_title">Notitie</string>
     <string name="verify_deposit_node_address_title">Knooppuntadres</string>
