+++ app/src/main/res/values-zh-rCN/strings.xml
@@ -384,7 +384,6 @@
     <string name="verify_swap_agree_allowance">我同意提供 ERC20 授权</string>
 
     <!-- Deposit Screen -->
-    <string name="deposit_screen_title">功能签名</string>
     <string name="deposit_screen_amount_title">金额</string>
     <string name="verify_deposit_memo_title">备注</string>
     <string name="verify_deposit_node_address_title">节点地址</string>
@@ -692,7 +691,7 @@
     <!-- Scan QR Screen -->
 
     <!-- Verify Transaction -->
-    <string name="verify_transaction_function_signature_title">功能签名</string>
+    <string name="verify_transaction_function_signature_title">函数签名</string>
     <string name="verify_transaction_function_inputs_title">函数参数</string>
     <string name="dapp_hero_unverified_function_title">未验证的功能</string>
     <string name="dapp_hero_unverified_function_subtitle">请在签名前查看下方详情</string>
