+++ app/src/main/res/values/strings.xml
@@ -296,7 +296,6 @@
     <string name="share_qr_screen_permission_required">File permission is required for share operations.</string>
     <string name="token_selection_search_hint">Search…</string>
     <string name="verify_swap_agree_allowance">I agree with providing ERC20 allowance</string>
-    <string name="deposit_screen_title">Function Signature</string>
     <string name="deposit_screen_amount_title">Amount</string>
     <string name="verify_deposit_memo_title">Memo</string>
     <string name="verify_deposit_node_address_title">Node Address</string>
@@ -1018,8 +1017,8 @@
     <string name="solana_instruction_number">Instruction %1$d</string>
     <string name="solana_program_id">Program ID: %1$s</string>
     <plurals name="solana_accounts_data">
-        <item quantity="one">Accounts: %1$d | Data Length: %2$d bytes</item>
-        <item quantity="other">Accounts: %1$d | Data Length: %2$d bytes</item>
+        <item quantity="one">Accounts: %1$d | Data length: %2$d bytes</item>
+        <item quantity="other">Accounts: %1$d | Data length: %2$d bytes</item>
     </plurals>
 
     <!-- Vault count plural -->
